Providing a linter configuration for project
Adding a .clang-format file with configuration of the code style.
Providing a script to check the format of source codes, which will
show a diff comparing current code format and expected code format for
files out of code style. Also providing a script for fixing the format
of the source code of the project.