Adding return code enum for BTLV Lib API functions
All BTLV lib functions now return a code from within a enum for succes,
error and warning codes. It adds more control and information about what
is happening internally in lib functions. It was designed in a way that
the users can ignore specific code and work only with the interpretation
of OK, error and warning depending if the returned code is negative,
which means error, greater than zero, meaning warning or zero, which
means OK.