Adding handler based on architecture for size_t
Parsing is faster when the memory is reinterpret. For those architecture
that uses big-endian, like the BER-TLV encoding, its benefical to just
interpret the block of memory holding length field as size_t and copy
this value. For those architectures that use little or pdp endian,
the handler correctly sort the bytes into memory before copying.